-Swapped roles of ' and non-' notation in S2, i.e exchanged:   --schubertCycle and schubertCycle'   --projectiveSpace and projectiveSpace'   --projectiveBundle and projectiveBundle'   --PP and PP' -Fixed documentation for schubertCycle and schubertCycle' to reflect this
This change was very desctrutive, documentation-wise.  It likely also broke
some test code.  Almost all of this (e.g. the test code) should be fixable by a find-and-replace,
but there are points when HTML tags, comments, and other things should be
modified.  Needs to be gone over with a fine-toothed comb.



git-svn-id: svn://svn.macaulay2.com/Macaulay2/workshops/Colorado-2010@12375 a8012f80-5f13-0410-8dc0-eb77bd940c0d